Added tags to events and severities page
