+++ src/main/javascript/trends.js
@@ -17,11 +17,6 @@ airframes.unshift("All Airframes");
 var index = airframes.indexOf("Garmin Flight Display");
 if (index !== -1) airframes.splice(index, 1);
 
-tagNames.unshift("All Tags");
-var tagIndex = tagNames.indexOf("Garmin Flight Display");
-if (tagIndex !== -1) tagNames.splice(tagIndex, 1);
-
-
 eventNames.sort();
 
 console.log(eventNames);
@@ -61,7 +56,6 @@ class TrendsPage extends React.Component {
         var date = new Date();
         this.state = {
             airframe : "All Airframes",
-            tagName: "All Tags",
             startYear : 2020,
             startMonth : 1,
             endYear : date.getFullYear(),
@@ -615,12 +609,6 @@ class TrendsPage extends React.Component {
         this.displayPlots(airframe);
     }
 
-    tagNameChange(tagName) {
-        this.setState({tagName});
-        this.displayPlots(tagName);
-    }
-
-
     render() {
         //console.log(systemIds);
 
@@ -654,9 +642,7 @@ class TrendsPage extends React.Component {
                                     updateEndYear={(newEndYear) => this.updateEndYear(newEndYear)}
                                     updateEndMonth={(newEndMonth) => this.updateEndMonth(newEndMonth)}
                                     exportCSV={() => this.exportCSV()}
-                                    tagNames={tagNames}
-                                    tagName={this.state.tagName}
-                                    tagNameChange={(tagName) => this.tagNameChange(tagName)}
+
 
                                 />
